+++ src/audio/PoliceRadio.cpp
@@ -176,7 +176,7 @@ cAudioManager::ServicePoliceRadio()
 }
 
 void
-cAudioManager::ServicePoliceRadioChannel(int32 wantedLevel)
+cAudioManager::ServicePoliceRadioChannel(uint8 wantedLevel)
 {
 	bool processed = false;
 	uint32 sample;
